Fixed up unicode environment variable issue (#156)
We improperly eagerly closed #154 with a separate issue. It turns out our injection of `sublime_terminal_path` in Python 2 (ST2) as a unicode string was causing issues as `Popen` only wants `str` strings. In this PR:

- Added normalization for environment from `unicode` to `str`

**Notes:**

We could move from `create_unicode_buffer` to `create_string_buffer` but I have a feeling that could undo existing patched regressions and this seems like it should work fine. Tested on ST2 on Windows and ST3 on Linux